# apps/api/src/domain.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Headers,
  HttpCode,
  Inject,
  Param,
  Post,
  Query,
  Req,
} from '@nestjs/common';
import {
  ApiBadRequestResponse,
  ApiBearerAuth,
  ApiBody,
  ApiConflictResponse,
  ApiCreatedResponse,
  ApiForbiddenResponse,
  ApiHeader,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiQuery,
  ApiTags,
  ApiUnauthorizedResponse,
} from '@nestjs/swagger';
import {
  accountIdentityMutationResponseSchema,
  accountIdentityProviderParameterSchema,
  accountOverviewSchema,
  accountRequestMutationResponseSchema,
  apiErrorEnvelopeSchema,
  apiPaginationQuerySchema,
  authenticatedContextSchema,
  businessReachOptionsSchema,
  businessCampaignDetailSchema,
  businessCampaignPageSchema,
  createAccountRequestSchema,
  createMissionApplicationRequestSchema,
  creatorMissionDetailSchema,
  creatorMissionPageSchema,
  creatorReachOverviewSchema,
  idempotencyKeySchema,
  linkAccountIdentityRequestSchema,
  missionApplicationResponseSchema,
  revokeAccountSessionRequestSchema,
  revokeAccountSessionResponseSchema,
  socialPlatformSchema,
  unlinkAccountIdentityRequestSchema,
} from '@local-missions/contracts';
import { z } from 'zod';

import { ApiProblem, validationProblem } from './api-errors.js';
import type { ContextualRequest } from './api-context.js';
import { DomainApiService } from './domain-api.service.js';
import { openApiSchema } from './openapi.js';

const publicIdParameterSchema = z.string().regex(/^cmp_[a-z0-9_]{8,100}$/);

@ApiTags('authenticated-v1')
@ApiBearerAuth('entra-bearer')
@ApiUnauthorizedResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
@ApiForbiddenResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
@Controller('v1')
export class DomainController {
  constructor(@Inject(DomainApiService) private readonly domain: DomainApiService) {}

  @ApiOperation({ summary: 'Resolve the current role and workspace context' })
  @ApiOkResponse({ schema: openApiSchema(authenticatedContextSchema) })
  @Get('me')
  async getMe(@Req() request: ContextualRequest) {
    return (await this.domain.authenticate(request)).context;
  }

  @ApiOperation({ summary: 'Read safe identity, session, and account-request state' })
  @ApiOkResponse({ schema: openApiSchema(accountOverviewSchema) })
  @Get('account')
  async getAccountOverview(@Req() request: ContextualRequest) {
    const principal = await this.domain.authenticate(request);
    return this.domain.getAccountOverview(principal);
  }

  @ApiOperation({ summary: 'Link a provider identity from verified provider control' })
  @ApiBody({ schema: openApiSchema(linkAccountIdentityRequestSchema) })
  @ApiCreatedResponse({ schema: openApiSchema(accountIdentityMutationResponseSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Post('account/identities')
  async linkAccountIdentity(@Body() input: unknown, @Req() request: ContextualRequest) {
    const body = linkAccountIdentityRequestSchema.safeParse(input);
    if (!body.success) throw validationProblem(body.error.issues, 'body');
    const principal = await this.domain.authenticate(request);
    return this.domain.linkAccountIdentity(
      principal,
      body.data,
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'Unlink one verified provider while retaining another method' })
  @ApiParam({ name: 'provider', schema: openApiSchema(accountIdentityProviderParameterSchema) })
  @ApiBody({ schema: openApiSchema(unlinkAccountIdentityRequestSchema) })
  @ApiOkResponse({ schema: openApiSchema(accountIdentityMutationResponseSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Delete('account/identities/:provider')
  async unlinkAccountIdentity(
    @Body() input: unknown,
    @Param('provider') provider: string,
    @Req() request: ContextualRequest,
  ) {
    const body = unlinkAccountIdentityRequestSchema.safeParse(input);
    if (!body.success) throw validationProblem(body.error.issues, 'body');
    const parsedProvider = accountIdentityProviderParameterSchema.safeParse(provider);
    if (!parsedProvider.success) {
      throw validationProblem(parsedProvider.error.issues, 'path.provider');
    }
    const principal = await this.domain.authenticate(request);
    return this.domain.unlinkAccountIdentity(
      principal,
      parsedProvider.data,
      body.data,
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'Revoke one active session owned by the current account' })
  @ApiBody({ schema: openApiSchema(revokeAccountSessionRequestSchema) })
  @ApiOkResponse({ schema: openApiSchema(revokeAccountSessionResponseSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @HttpCode(200)
  @Post('account/logout')
  async revokeAccountSession(@Body() input: unknown, @Req() request: ContextualRequest) {
    const body = revokeAccountSessionRequestSchema.safeParse(input);
    if (!body.success) throw validationProblem(body.error.issues, 'body');
    const principal = await this.domain.authenticate(request);
    return this.domain.revokeAccountSession(
      principal,
      body.data.sessionPublicId,
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'Request an account export or recent-authenticated deletion' })
  @ApiBody({ schema: openApiSchema(createAccountRequestSchema) })
  @ApiCreatedResponse({ schema: openApiSchema(accountRequestMutationResponseSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Post('account/requests')
  async createAccountRequest(@Body() input: unknown, @Req() request: ContextualRequest) {
    const body = createAccountRequestSchema.safeParse(input);
    if (!body.success) throw validationProblem(body.error.issues, 'body');
    const principal = await this.domain.authenticate(request);
    return this.domain.createAccountRequest(
      principal,
      body.data,
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'Read optional per-platform Reach status for the current Creator' })
  @ApiOkResponse({ schema: openApiSchema(creatorReachOverviewSchema) })
  @Get('creator/reach')
  async getCreatorReach(@Req() request: ContextualRequest) {
    const principal = await this.domain.authenticate(request);
    return this.domain.getCreatorReach(principal);
  }

  @ApiOperation({ summary: 'Grant optional read-only Reach analytics consent for one platform' })
  @ApiParam({ name: 'platform', schema: openApiSchema(socialPlatformSchema) })
  @ApiOkResponse({ schema: openApiSchema(creatorReachOverviewSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @HttpCode(200)
  @Post('creator/reach/:platform/consent')
  async grantCreatorReachConsent(
    @Param('platform') platform: string,
    @Req() request: ContextualRequest,
  ) {
    const principal = await this.domain.authenticate(request);
    return this.domain.grantCreatorReachConsent(
      principal,
      this.parseSocialPlatform(platform),
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'Revoke optional Reach analytics consent for one platform' })
  @ApiParam({ name: 'platform', schema: openApiSchema(socialPlatformSchema) })
  @ApiOkResponse({ schema: openApiSchema(creatorReachOverviewSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Delete('creator/reach/:platform/consent')
  async revokeCreatorReachConsent(
    @Param('platform') platform: string,
    @Req() request: ContextualRequest,
  ) {
    const principal = await this.domain.authenticate(request);
    return this.domain.revokeCreatorReachConsent(
      principal,
      this.parseSocialPlatform(platform),
      request.apiContext?.correlationId ?? 'request_context_unavailable',
    );
  }

  @ApiOperation({ summary: 'List published Community missions with available capacity' })
  @ApiQuery({ name: 'cursor', required: false, type: String })
  @ApiQuery({ maximum: 100, minimum: 1, name: 'limit', required: false, type: Number })
  @ApiOkResponse({ schema: openApiSchema(creatorMissionPageSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Get('creator/missions')
  async listCreatorMissions(@Query() input: unknown, @Req() request: ContextualRequest) {
    const query = apiPaginationQuerySchema.safeParse(input);
    if (!query.success) throw validationProblem(query.error.issues, 'query');
    const principal = await this.domain.authenticate(request);
    return this.domain.listCreatorMissions(principal, query.data.limit, query.data.cursor);
  }

  @ApiOperation({ summary: 'Read an available Community mission contract' })
  @ApiParam({ name: 'campaignPublicId', type: String })
  @ApiOkResponse({ schema: openApiSchema(creatorMissionDetailSchema) })
  @ApiNotFoundResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Get('creator/missions/:campaignPublicId')
  async getCreatorMission(
    @Param('campaignPublicId') campaignPublicId: string,
    @Req() request: ContextualRequest,
  ) {
    const parsedId = this.parseCampaignId(campaignPublicId);
    const principal = await this.domain.authenticate(request);
    return this.domain.getCreatorMission(principal, parsedId);
  }

  @ApiOperation({ summary: 'Apply once for an available Community Slot' })
  @ApiBody({ schema: openApiSchema(createMissionApplicationRequestSchema) })
  @ApiHeader({ name: 'idempotency-key', required: true })
  @ApiParam({ name: 'campaignPublicId', type: String })
  @ApiCreatedResponse({ schema: openApiSchema(missionApplicationResponseSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @ApiConflictResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @ApiNotFoundResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Post('creator/missions/:campaignPublicId/applications')
  async applyForCreatorMission(
    @Body() input: unknown,
    @Param('campaignPublicId') campaignPublicId: string,
    @Headers('idempotency-key') idempotencyKey: string | undefined,
    @Req() request: ContextualRequest,
  ) {
    const body = createMissionApplicationRequestSchema.safeParse(input);
    if (!body.success) throw validationProblem(body.error.issues, 'body');
    const key = idempotencyKeySchema.safeParse(idempotencyKey);
    if (!key.success) {
      throw new ApiProblem(
        'IDEMPOTENCY_KEY_REQUIRED',
        'A valid idempotency key is required.',
        400,
        [{ code: 'INVALID_FORMAT', path: 'header.idempotency-key' }],
      );
    }
    const principal = await this.domain.authenticate(request);
    return this.domain.applyForMission({
      campaignPublicId: this.parseCampaignId(campaignPublicId),
      correlationId: request.apiContext?.correlationId ?? 'request_context_unavailable',
      idempotencyKey: key.data,
      principal,
      publicId: body.data.publicId,
    });
  }

  @ApiOperation({ summary: 'List campaigns in the active Business workspace' })
  @ApiQuery({ name: 'cursor', required: false, type: String })
  @ApiQuery({ maximum: 100, minimum: 1, name: 'limit', required: false, type: Number })
  @ApiOkResponse({ schema: openApiSchema(businessCampaignPageSchema) })
  @ApiBadRequestResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Get('business/campaigns')
  async listBusinessCampaigns(@Query() input: unknown, @Req() request: ContextualRequest) {
    const query = apiPaginationQuerySchema.safeParse(input);
    if (!query.success) throw validationProblem(query.error.issues, 'query');
    const principal = await this.domain.authenticate(request);
    return this.domain.listBusinessCampaigns(principal, query.data.limit, query.data.cursor);
  }

  @ApiOperation({ summary: 'Read a campaign in the active Business workspace' })
  @ApiParam({ name: 'campaignPublicId', type: String })
  @ApiOkResponse({ schema: openApiSchema(businessCampaignDetailSchema) })
  @ApiNotFoundResponse({ schema: openApiSchema(apiErrorEnvelopeSchema) })
  @Get('business/campaigns/:campaignPublicId')
  async getBusinessCampaign(
    @Param('campaignPublicId') campaignPublicId: string,
    @Req() request: ContextualRequest,
  ) {
    const principal = await this.domain.authenticate(request);
    return this.domain.getBusinessCampaign(principal, this.parseCampaignId(campaignPublicId));
  }

  @ApiOperation({ summary: 'Read fixed Reach packages and per-platform availability' })
  @ApiOkResponse({ schema: openApiSchema(businessReachOptionsSchema) })
  @Get('business/reach-options')
  async getBusinessReachOptions(@Req() request: ContextualRequest) {
    const principal = await this.domain.authenticate(request);
    return this.domain.getBusinessReachOptions(principal);
  }

  private parseCampaignId(value: string): string {
    const parsed = publicIdParameterSchema.safeParse(value);
    if (!parsed.success) throw validationProblem(parsed.error.issues, 'path.campaignPublicId');
    return parsed.data;
  }

  private parseSocialPlatform(value: string) {
    const parsed = socialPlatformSchema.safeParse(value);
    if (!parsed.success) throw validationProblem(parsed.error.issues, 'path.platform');
    return parsed.data;
  }
}